requirements.txt
Flask==1.1.1
Flask-PyMongo==2.3.0
pymongo==3.9.0
pytest-flask==0.15.0
